requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-timeout==4.0.2
attrs==21.4.0
avro==1.10.2
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
confluent-kafka==1.9.2
cryptography==36.0.2
cx-Oracle==8.3.0
decorator==5.1.1
Deprecated==1.2.13
docker==6.0.0
elastic-transport==8.4.0
elasticsearch==8.6.1
execnet==1.9.0
frozenlist==1.3.0
greenlet==1.1.2
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.1
jsonpath-ng==1.5.3
jsonschema==4.4.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.1
packaging==21.3
pandas==1.4.3
paramiko==2.10.3
pluggy==1.0.0
ply==3.11
protobuf==4.21.12
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyhcl==0.4.4
pyignite==0.6.0
PyJWT==2.3.0
pymongo==4.1.1
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
pytest-csv==3.0.0
pytest-forked==1.4.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-graphql-client==0.4.3
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
requests-futures==1.0.0
rstr==3.0.0
six==1.16.0
SQLAlchemy==1.4.31
tabulate==0.8.9
thrift==0.16.0
thriftpy2==0.4.16
toml==0.10.2
urllib3==1.26.8
websocket-client==1.2.3
websockets==10.2
wrapt==1.13.3
yarl==1.7.2